feature: Reverse display order for "Open Tabs Above" option
This is a new take on the "Open Tabs Above" option. Instead of inserting tabs before the current tab, only the display order is changed. The effect is that tabs are
automatically inserted in the correct position by Firefox. Also Ctrl+Tab works the same way as without TabCenter (by going to the new tab).

Fixes: #709
BREAKING CHANGE: Ctrl+Tab now selects tabs in the correct order.