Scroll to open tab correctly after exiting the sidebar
The `skipNextScroll` workaround for #730 is only necessary when the
sidebar is set to automatically shrink, otherwise in some cases we end
up with `skipNextScroll` set to true after exiting the sidebar. This
means that the sidebar won't scroll to display the next tab opened, see
 #851 for more details.

This commit fixes this by only setting `skipNextScroll` to true if we
are in shrink mode.

Fixes #851