Restore browser-bottombox location before customize
This fixes compatibility issue with Toolbar Position Changer add-on that
enables to move toolbars. TPC uses browser-bottombox as "bottom" toolbox
container for toolbars. Tab Center  moves bottombox to different location
which is not used by Firefox when entering to the Customize mode. This fix
restores the original position of bottombox in the beforeListener.