fix #897: animated visuallyselected indicator
Implementations of the left blue border (for selected tabs) using a background-image or box-shadow + transition on the tab itself regularly failed due to CSS conflicts, as in #897.

I cleaned up code for the box-shadow implementation and some leftover from a previous implementation (background-image I think), and reimplemented the indicator using a XUL element.

The gradient is now smoother and composed of a single color with a darkening gradient on top, which allows changing that color using a CSS variable, e.g. for theming or maybe picking up a Container Tab color.